Add overwrite option in evaluate squad script
Add 8 and 20 epochs experiments

+++ evaluate_squad.sh
@@ -2,6 +2,7 @@
 
 
 data_dir=${1:-results}; shift
+overwrite=${1:-0}; shift
 correct_thr=${1:-0}; shift
 predictions=$(ls -d $data_dir/*/ | cut -f2 -d'/')
 
@@ -16,6 +17,10 @@ for prediction in ${predictions[@]}; do
       source $exp
       output=${prediction_data}/results.json
       # get new threshold
+      if [[ -f $output && $overwrite -eq 0 ]]; then 
+        echo "Skipping $(basename ${prediction_data}), use overwrite=1 to force it anyway."
+        continue
+      fi
       echo "$(basename $PREDICT_FILE) $(basename ${prediction_data})/predictions.json --na-prob-file $(basename ${prediction_data})/null_odds.json"
       new_thrs=$(python $eval_script $PREDICT_FILE ${prediction_data}/predictions.json --na-prob-file ${prediction_data}/null_odds.json | tee $output | grep "best_f1_thresh" | awk '{print $2}')
       if [[ $correct_thr -ne 0 ]]; then
